fix: Fix type assertions and event dispatching in Stablecoin tests
- Change amount parameter type from int to float in CollateralLocked event test
- Update CollateralService method signatures to return float instead of int
- Fix DemoStablecoinService to dispatch events with float amounts
- Replace event assertions with database assertions in DemoStablecoinServiceTest
  (Events within DB transactions are not visible to Event::fake())

+++ app/Domain/Stablecoin/Services/DemoStablecoinService.php
@@ -78,7 +78,7 @@ class DemoStablecoinService
                 position_uuid: $positionId,
                 account_uuid: $accountId,
                 stablecoin_code: $stablecoinId,
-                amount: (int) $amount,  // Use the original amount for the event
+                amount: (float) $amount,  // Use the original amount for the event
                 metadata: ['collateral' => $collateral]
             ));
 
@@ -149,7 +149,7 @@ class DemoStablecoinService
                 position_uuid: $position->uuid,
                 account_uuid: $accountId,
                 stablecoin_code: $stablecoinId,
-                amount: (int) $amount,
+                amount: (float) $amount,
                 metadata: ['collateral_returned' => $collateralToReturn]
             ));
